+++ src/search/landmarks/landmark_graph.h
@@ -72,7 +72,7 @@ public:
 
     bool is_true_in_state(const State &state) const {
         if (disjunctive) {
-            for (unsigned int i = 0; i < vars.size(); ++i) {
+            for (size_t i = 0; i < vars.size(); ++i) {
                 if (state[vars[i]] == vals[i]) {
                     return true;
                 }
